Refactoring data
