+++ base/shared/src/main/scala/scalaz/ct/invariantFunctor.scala
