--- base/shared/src/main/scala/scalaz/data/DisjunctionFunctions.scala
@@ -1,12 +1,0 @@
-package scalaz
-package data
-
-trait DisjunctionFunctions {
-  @inline def left[L, R](value: L): Disjunction[L, R]  = -\/(value)
-  @inline def right[L, R](value: R): Disjunction[L, R] = \/-(value)
-
-  def either[A, B, C](ac: A => C)(bc: B => C): A \/ B => C = _ match {
-    case -\/(l) => ac(l)
-    case \/-(r) => bc(r)
-  }
-}
