--- base/shared/src/main/scala/scalaz/data/DisjunctionSyntax.scala
@@ -1,13 +1,0 @@
-package scalaz
-package data
-
-trait DisjunctionSyntax {
-  implicit final class ToDisjunctionOps[A](a: A) {
-    def left[B]: A \/ B  = -\/(a)
-    def right[B]: B \/ A = \/-(a)
-  }
-
-  implicit final class EitherAsDisjunction[A, B](ab: Either[A, B]) {
-    def asDisjunction: A \/ B = Disjunction.fromEither(ab)
-  }
-}
