--- base/shared/src/main/scala/scalaz/data/DownStar.scala
@@ -1,6 +1,0 @@
-package scalaz
-package data
-
-final case class DownStar[F[_], A, B](run: F[A] => B) extends AnyVal
-
-object DownStar extends DownStarInstances
