--- base/shared/src/main/scala/scalaz/data/Identity.scala
@@ -1,4 +1,0 @@
-package scalaz
-package data
-
-final case class Identity[A](run: A) extends AnyVal
