+++ base/shared/src/main/scala/scalaz/data/aCatenable1.scala
