+++ base/shared/src/main/scala/scalaz/data/aMaybe2.scala
