+++ base/shared/src/main/scala/scalaz/data/as.scala
@@ -1,6 +1,7 @@
 package scalaz
 package data
 
+import scala.language.implicitConversions
 /**
  * Liskov substitutability: A better `<:<`.
  *
@@ -266,4 +267,30 @@ object As extends AsInstances {
    */
   def unsafeForce[A, B]: A <~< B =
     refl[Any].asInstanceOf[A <~< B]
+}
+
+trait AsInstances {
+
+  /**We can witness equality by using it to convert between types */
+  implicit def witness[A, B](lt: A <~< B): A => B = lt(_)
+}
+
+trait AsSyntax {
+  implicit final class ToAsOps[A, B](val ab: As[A, B]) {
+    def liftCvF[F[_]](implicit F: IsCovariant[F]): F[A] As F[B] =
+      F.liftLiskov(ab)
+
+    def liftCtF[F[_]](implicit F: IsContravariant[F]): F[B] As F[A] =
+      F.liftLiskov(ab)
+
+    def substCvF[F[_]](fa: F[A])(implicit F: IsCovariant[F]): F[B] = {
+      type f[+x] = x
+      F.substCv[f, A, B](fa)(ab)
+    }
+
+    def substCtF[F[_]](fb: F[B])(implicit F: IsContravariant[F]): F[A] = {
+      type f[+x] = x
+      F.substCv[f, A, B](fb)(ab)
+    }
+  }
 }
