+++ base/shared/src/main/scala/scalaz/data/const.scala
@@ -6,6 +6,12 @@ import scalaz.core.EqClass
 import scalaz.ct._
 import scalaz.debug.DebugClass
 
+final case class Const[A, B](getConst: A) {
+  def retag[C]: Const[A, C] = this.asInstanceOf[Const[A, C]]
+}
+
+object Const extends ConstInstances
+
 trait ConstInstances {
   implicit def constTraverse[R]: Traversable[Const[R, ?]] =
     instanceOf(
