+++ base/shared/src/main/scala/scalaz/data/downStar.scala
@@ -1,6 +1,10 @@
 package scalaz
 package data
 
+final case class DownStar[F[_], A, B](run: F[A] => B) extends AnyVal
+
+object DownStar extends DownStarInstances
+
 import scalaz.ct.ProfunctorClass
 
 trait DownStarInstances {
