+++ base/shared/src/main/scala/scalaz/data/endo.scala
