+++ base/shared/src/main/scala/scalaz/data/forall.scala
