+++ base/shared/src/main/scala/scalaz/data/forall2.scala
