+++ base/shared/src/main/scala/scalaz/data/forget.scala
@@ -3,6 +3,10 @@ package data
 
 import ct.StrongClass
 
+final case class Forget[A, B, C](forget: B => A) {
+  def retag[D]: Forget[A, B, D] = this.asInstanceOf[Forget[A, B, D]]
+}
+
 trait ForgetInstances { self =>
   implicit def strong[A]: Strong[Forget[A, ?, ?]] =
     instanceOf(new StrongClass[Forget[A, ?, ?]] {
