+++ base/shared/src/main/scala/scalaz/data/is.scala
