+++ base/shared/src/main/scala/scalaz/types/isContravariant.scala
