+++ base/shared/src/main/scala/scalaz/types/isCovariant.scala
